refactor: Move username components out of auth
Rename and relocate username components from components/auth/username to components/username. Update all imports (components, plugin, and tests) to the new paths and adjust public exports in plugins.ts. This flattens the component structure for username-related pieces and keeps tests and the username plugin in sync.
